Added projectile attack

+++ Assets/Scripts/Projectiles/Projectile.cs
@@ -6,7 +6,7 @@ public abstract class Projectile : MonoBehaviour
 {
     public GameObject Enemy { get; set; }
     public GameObject CurrentEnemy { get; set; }
-    protected float Damage { get; set; }
+    protected int Damage { get; set; }
 
     public void SetEnemy(GameObject enemy)
     {
